Endpoint to duplicate registrations AB#29123 (#5552)
AB#29123 Endpoint to duplicate registrations
