+++ services/121-service/src/scripts/sql/mock-messages.sql
@@ -1,6 +1,5 @@
  INSERT INTO "121-service"."twilio_message"
         (
-          id,
           created,
           "accountSid",
           body,
@@ -18,10 +17,6 @@
           "errorMessage"
         )
         SELECT
-          id + (
-            SELECT count(id)
-            FROM "121-service"."twilio_message"
-          ),
           created + random() * (timestamp '2023-01-01 00:00:00' -
                    timestamp '2023-01-02 00:00:00'),
           "accountSid",
